refactor: refactored calculate tab (isolated concerns, reused existing functions and classes were possible)
